+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Moved `-[NSString randomStringWithLength:]` to `ESRandomString()` helper
 - Moved `-[NSData randomDataWithLength:]` to `ESRandomData()` helper
 - Moved `-[UIColor randomColor]` to `ESRandomColor()` helper
+- Make `ESIsPadUI()`, `ESIsPadDevice()`, `ESIsRetinaScreen()` etc public for tvOS
 
 ## 3.16.1 (2019-07-25)